requirements.txt
whisper
Pillow